requirements.txt
alembic==1.13.2
APScheduler==3.10.4
asgiref==3.8.1
autopep8==2.0.4
bidict==0.23.1
black==23.11.0
bleach==6.1.0
blinker==1.8.2
cffi==1.16.0
click==8.1.7
colorama==0.4.6
dnspython==2.6.1
Flask==3.0.3
Flask-Admin==1.6.1
Flask-Cors==4.0.1
Flask-JWT-Extended==4.6.0
Flask-Login==0.6.3
Flask-Migrate==4.0.7
Flask-PyMongo==2.3.0
Flask-SocketIO==5.3.6
Flask-SQLAlchemy==3.1.1
gevent==24.2.1
gevent-websocket==0.10.1
greenlet==3.0.3
gunicorn==22.0.0
h11==0.14.0
itsdangerous==2.2.0
Jinja2==3.1.4
Mako==1.3.5
MarkupSafe==2.1.5
mypy-extensions==1.0.0
packaging==23.2
pathspec==0.11.2
platformdirs==3.11.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
pycodestyle==2.11.1
pycparser==2.22
PyJWT==2.8.0
pymongo==4.8.0
python-dotenv==1.0.1
python-engineio==4.9.1
python-socketio==5.11.3
pytz==2024.1
simple-websocket==1.0.0
six==1.16.0
SQLAlchemy==2.0.30
sqlparse==0.4.4
typing_extensions==4.12.2
tzdata==2024.1
tzlocal==5.2
waitress==3.0.0
webencodings==0.5.1
Werkzeug==3.0.3
wsproto==1.2.0
WTForms==3.1.2
zope.event==5.0
zope.interface==6.4.post2
